# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "assignment-5-customer-data",
    "version": "1.0.0",
    "description": "API for managing customers",
    "contact": {}
  },
  "servers": [],
  "paths": {
    "/customers": {
      "get": {
        "summary": "Get all customers",
        "description": "Retrieves a list of all customers with their associated addresses, contacts, and orders.",
        "operationId": "getAllCustomers",
        "tags": ["Customers"],
        "responses": {
          "200": {
            "description": "Successfully retrieved customers list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "cs_id": {
                        "type": "integer",
                        "description": "Customer ID"
                      },
                      "cs_fullName": {
                        "type": "string",
                        "description": "Full name of the customer"
                      },
                      "cs_email": {
                        "type": "string",
                        "description": "Email address of the customer"
                      },
                      "cs_gender": {
                        "type": "string",
                        "enum": ["Male", "Female"],
                        "description": "Gender of the customer"
                      },
                      "cs_birthDate": {
                        "type": "string",
                        "format": "date",
                        "description": "Birth date of the customer"
                      },
                      "addresses": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "adrs_id": {
                              "type": "integer",
                              "description": "Address ID"
                            },
                            "adrs_street": {
                              "type": "string",
                              "description": "Street address"
                            },
                            "adrs_city": {
                              "type": "string",
                              "description": "City"
                            },
                            "adrs_state": {
                              "type": "string",
                              "description": "State"
                            },
                            "adrs_postalCode": {
                              "type": "string",
                              "description": "Postal code"
                            }
                          }
                        }
                      },
                      "contacts": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "cnt_id": {
                              "type": "integer",
                              "description": "Contact ID"
                            },
                            "cnt_phone": {
                              "type": "string",
                              "description": "Phone number"
                            },
                            "cnt_fax": {
                              "type": "string",
                              "description": "Fax number"
                            }
                          }
                        }
                      },
                      "orders": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "ord_id": {
                              "type": "integer",
                              "description": "Order ID"
                            },
                            "ord_date": {
                              "type": "string",
                              "format": "date-time",
                              "description": "Order date"
                            },
                            "ord_status": {
                              "type": "string",
                              "enum": ["pending", "completed", "cancelled"],
                              "description": "Order status"
                            },
                            "ord_totalAmount": {
                              "type": "number",
                              "format": "decimal",
                              "description": "Total amount of the order"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customers not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customers not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "description": "Create a new customer with full name, email, gender, and birth date.",
        "operationId": "createCustomer",
        "tags": ["Customers"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cs_fullName": {
                    "type": "string",
                    "description": "Full name of the customer",
                    "example": "John Doe"
                  },
                  "cs_email": {
                    "type": "string",
                    "description": "Email address of the customer",
                    "example": "john.doe@example.com"
                  },
                  "cs_gender": {
                    "type": "string",
                    "description": "Gender of the customer",
                    "enum": ["Male", "Female", "Other"],
                    "example": "Male"
                  },
                  "cs_birthDate": {
                    "type": "string",
                    "format": "date",
                    "description": "Birth date of the customer",
                    "example": "1990-01-01"
                  }
                },
                "required": [
                  "cs_fullName",
                  "cs_email",
                  "cs_gender",
                  "cs_birthDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer created"
                    },
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "cs_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "cs_fullName": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "cs_email": {
                          "type": "string",
                          "example": "john.doe@example.com"
                        },
                        "cs_gender": {
                          "type": "string",
                          "example": "Male"
                        },
                        "cs_birthDate": {
                          "type": "string",
                          "format": "date",
                          "example": "1990-01-01"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields are required"
                    },
                    "status": {
                      "type": "integer",
                      "example": 400
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email already exists"
                    },
                    "status": {
                      "type": "integer",
                      "example": 409
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "status": {
                      "type": "integer",
                      "example": 500
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "summary": "Get customer by ID",
        "description": "Retrieves a customer by their unique ID with their associated addresses, contacts, and orders.",
        "operationId": "getCustomerById",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the customer to retrieve",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "cs_id": {
                      "type": "integer",
                      "description": "Customer ID"
                    },
                    "cs_fullName": {
                      "type": "string",
                      "description": "Full name of the customer"
                    },
                    "cs_email": {
                      "type": "string",
                      "description": "Email address of the customer"
                    },
                    "cs_gender": {
                      "type": "string",
                      "enum": ["Male", "Female"],
                      "description": "Gender of the customer"
                    },
                    "cs_birthDate": {
                      "type": "string",
                      "format": "date",
                      "description": "Birth date of the customer"
                    },
                    "addresses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "adrs_id": {
                            "type": "integer",
                            "description": "Address ID"
                          },
                          "adrs_street": {
                            "type": "string",
                            "description": "Street address"
                          },
                          "adrs_city": {
                            "type": "string",
                            "description": "City"
                          },
                          "adrs_state": {
                            "type": "string",
                            "description": "State"
                          },
                          "adrs_postalCode": {
                            "type": "string",
                            "description": "Postal code"
                          }
                        }
                      }
                    },
                    "contacts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "cnt_id": {
                            "type": "integer",
                            "description": "Contact ID"
                          },
                          "cnt_phone": {
                            "type": "string",
                            "description": "Phone number"
                          },
                          "cnt_fax": {
                            "type": "string",
                            "description": "Fax number"
                          }
                        }
                      }
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "ord_id": {
                            "type": "integer",
                            "description": "Order ID"
                          },
                          "ord_date": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Order date"
                          },
                          "ord_status": {
                            "type": "string",
                            "enum": ["pending", "completed", "cancelled"],
                            "description": "Order status"
                          },
                          "ord_totalAmount": {
                            "type": "number",
                            "format": "decimal",
                            "description": "Total amount of the order"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update customer by ID",
        "description": "Updates the details of a customer by their unique ID. All fields (`cs_fullName`, `cs_email`, `cs_gender`, `cs_birthDate`) are required.",
        "operationId": "updateCustomer",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the customer to update",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cs_fullName": {
                    "type": "string",
                    "description": "Full name of the customer"
                  },
                  "cs_email": {
                    "type": "string",
                    "description": "Email address of the customer"
                  },
                  "cs_gender": {
                    "type": "string",
                    "enum": ["Male", "Female"],
                    "description": "Gender of the customer"
                  },
                  "cs_birthDate": {
                    "type": "string",
                    "format": "date",
                    "description": "Birth date of the customer"
                  }
                },
                "required": [
                  "cs_fullName",
                  "cs_email",
                  "cs_gender",
                  "cs_birthDate"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer updated"
                    },
                    "previousDataCustomer": {
                      "type": "object",
                      "properties": {
                        "cs_id": {
                          "type": "integer",
                          "description": "Customer ID"
                        },
                        "cs_fullName": {
                          "type": "string",
                          "description": "Full name of the customer"
                        },
                        "cs_email": {
                          "type": "string",
                          "description": "Email address of the customer"
                        },
                        "cs_gender": {
                          "type": "string",
                          "enum": ["Male", "Female"],
                          "description": "Gender of the customer"
                        },
                        "cs_birthDate": {
                          "type": "string",
                          "format": "date",
                          "description": "Birth date of the customer"
                        }
                      }
                    },
                    "updatedCustomer": {
                      "type": "integer",
                      "description": "Number of rows updated"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request, required fields are missing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields are required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete customer by ID",
        "description": "Deletes a customer by their unique ID. If the customer does not exist, a 404 error will be returned.",
        "operationId": "deleteCustomer",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the customer to delete",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted the customer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}